Add command update_own_status
That accepts status type of validation or
puppet apply as well as the result.

The result is intended to be the exit code
from either puppet with detailed-exit-codes
or from runparts when it runs the local
nagios checks.

for Puppet: 1,4,6 map to failure, everything
else to success

for Validation: 0 maps to success, everything
else failure.