Fix bug in get_ip_of_node
Previously, an exception was being raised if the
first node from the list did not match the name
of the node whose ip address was being retrieve.

This only happened to work b/c the node name etcd
was the first alphabetically. When I added a contrail
node, it cased this code to break.