update_own_info should delete previous versions
we were relying on a ttl from etcd to ensure that
only the lastest version was ever written into the
kv store.

consul does not support ttls on the kv store, so we
need to ensure that we clean out old versions so that
check_sinlge_version and friends will actually work.

it is worth noting that this implementation is likely
temporary and that version might wind up being some
kind of health check instead. for that reason, the
potentially performance issues are not that big of
a deal.